pyproject.toml
[project]
name = "emoguard-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deadcode>=2.4.1",
    "django>=5.2.7",
    "django-ninja>=1.4.5",
    "django-ninja-extra>=0.30.2",
    "django-ninja-jwt>=5.4.0",
    "django-summernote>=0.8.20.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "pycryptodome",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.14.3",
    "uvicorn>=0.38.0",
]

[tool.deadcode]
exclude = [".venv", "venv", "__pycache__", "*.pyc", "migrations"]
ignore-names = ["*TestCase", "*Mixin", "test_*"]
